fix: 프로젝트 분석 querydsl 오류 수정

+++ src/main/java/org/causwengteam13/issuetrackerserver/domain/project/result/AnalyzeProjectByMemberResult.java
@@ -19,7 +19,7 @@ public record AnalyzeProjectByMemberResult(
 		@Builder
 		public record IssueStatistics(
 			IssueStatus status,
-			int count
+			Long count
 		) {
 		}
 	}
